Move token finding logic to DefaultVaultTokenLoader
Allows for easier tests and independent of Intellij and Vault communication
logic.
