+++ src/main/java/com/premiumminds/datagrip/vault/VaultDatabaseAuthProvider.java
@@ -24,6 +24,7 @@ import com.intellij.database.dataSource.DatabaseConnectionPoint;
 import com.intellij.openapi.diagnostic.Logger;
 import com.intellij.openapi.project.Project;
 import com.premiumminds.datagrip.vault.client.Credentials;
+import com.premiumminds.datagrip.vault.client.DefaultVaultTokenLoader;
 import com.premiumminds.datagrip.vault.client.VaultClient;
 import org.jetbrains.annotations.Nls;
 import org.jetbrains.annotations.NonNls;
@@ -39,12 +40,8 @@ public class VaultDatabaseAuthProvider implements DatabaseAuthProvider {
     public static final String PROP_TOKEN_FILE = "vault_token_file";
     private static final String ENV_VAULT_AGENT_ADDR = "VAULT_AGENT_ADDR";
     private static final String ENV_VAULT_ADDR = "VAULT_ADDR";
-    private static final String ENV_VAULT_CONFIG_PATH = "VAULT_CONFIG_PATH";
-    private static final String DEFAULT_VAULT_CONFIG_FILE = ".vault";
-    private static final String DEFAULT_VAULT_TOKEN_FILE = ".vault-token";
     private static final String ERROR_VAULT_ADDRESS_NOT_DEFINED = "Vault address not defined";
     private static final String ERROR_VAULT_SECRET_NOT_DEFINED = "Vault secret not defined";
-    private static final String ERROR_VAULT_TOKEN_NOT_DEFINED = "Vault token not defined";
 
     private static final VaultClient  vaultClient = new VaultClient();
 
@@ -75,16 +72,20 @@ public class VaultDatabaseAuthProvider implements DatabaseAuthProvider {
         logger.info("Address used: " + address);
         logger.info("Secret used: " + secret);
 
+        DefaultVaultTokenLoader vaultTokenLoader = new DefaultVaultTokenLoader(
+                Optional.ofNullable(protoConnection.getConnectionPoint().getAdditionalProperty(PROP_TOKEN_FILE)).map(Path::of),
+                address
+        );
+
         CacheKey key = new CacheKey(address, secret);
         Credentials value = secretsCache.compute(key, (k,v) -> {
             try {
-                final var token = getToken(protoConnection, address);
                 if (v == null) {
-                    return vaultClient.getCredentials(address, Optional.empty(), token, secret);
+                    return vaultClient.getCredentials(address, vaultTokenLoader, Optional.empty(), secret);
                 } else {
-                    final var lease = vaultClient.getLease(address, Optional.empty(), token, v.leaseId());
+                    final var lease = vaultClient.getLease(address, vaultTokenLoader, Optional.empty(), v.leaseId());
                     if (lease.isEmpty()) {
-                        return vaultClient.getCredentials(address, Optional.empty(), token, secret);
+                        return vaultClient.getCredentials(address, vaultTokenLoader, Optional.empty(), secret);
                     }
                 }
                 return v;
@@ -129,100 +130,5 @@ public class VaultDatabaseAuthProvider implements DatabaseAuthProvider {
             return secret;
         }
         throw new RuntimeException(ERROR_VAULT_SECRET_NOT_DEFINED);
-    }
-
-
-    private String getToken(ProtoConnection protoConnection, String vaultAddress) throws IOException, InterruptedException {
-
-        final var tokenFile = protoConnection.getConnectionPoint().getAdditionalProperty(PROP_TOKEN_FILE);
-        if (tokenFile != null && !tokenFile.isBlank()) {
-            final var path = Paths.get(tokenFile);
-            if (path.toFile().exists()){
-                return Files.readString(path);
-            }
-        }
-
-        final var vaultConfigFile = getConfigFile();
-        if (vaultConfigFile.toFile().exists()){
-            final String token = getTokenFromVaultTokenHelper(vaultConfigFile, vaultAddress);
-            if (token != null){
-                return token;
-            }
-        }
-        final var defaultTokenFilePath = Paths.get(System.getProperty("user.home"), DEFAULT_VAULT_TOKEN_FILE);
-        if (defaultTokenFilePath.toFile().exists()){
-            return Files.readString(defaultTokenFilePath);
-        }
-
-        throw new RuntimeException(ERROR_VAULT_TOKEN_NOT_DEFINED);
-    }
-
-    private Path getConfigFile(){
-        Path vaultConfigPath = Paths.get(System.getProperty("user.home"), DEFAULT_VAULT_CONFIG_FILE) ;
-
-        final String vaultConfigPathEnv = System.getenv(ENV_VAULT_CONFIG_PATH);
-        if (vaultConfigPathEnv != null && !vaultConfigPathEnv.isBlank()){
-            vaultConfigPath = Paths.get(vaultConfigPathEnv );
-        }
-
-        return vaultConfigPath;
-    }
-
-    private String getTokenFromVaultTokenHelper(Path configFile, String vaultAddress)
-            throws IOException, InterruptedException
-    {
-        final Gson gson = new Gson();
-
-        try (FileReader fileReader = new FileReader(configFile.toFile())) {
-            final VaultConfig config = gson.fromJson(fileReader, VaultConfig.class);
-
-            if (config.tokenHelper != null && !config.tokenHelper.isBlank()){
-                final ProcessBuilder processBuilder = new ProcessBuilder();
-                processBuilder.environment().putIfAbsent(ENV_VAULT_ADDR, vaultAddress);
-                final Process process = processBuilder
-                        .command(config.tokenHelper, "get")
-                        .start();
-
-                final StreamGobbler streamGobblerErr = new StreamGobbler(process.getErrorStream());
-                final StreamGobbler streamGobblerOut = new StreamGobbler(process.getInputStream());
-
-                streamGobblerErr.start();
-                streamGobblerOut.start();
-
-                if (!process.waitFor(10, TimeUnit.SECONDS)){
-                    throw new RuntimeException("Failure running Vault Token Helper: " + config.tokenHelper + ", took too long to respond.");
-                }
-
-                streamGobblerOut.join();
-                streamGobblerErr.join();
-
-                if (streamGobblerErr.output != null && !streamGobblerErr.output.isBlank()){
-                    throw new RuntimeException("Failure running Vault Token Helper: " + config.tokenHelper + ": " + streamGobblerErr.output);
-                }
-                return streamGobblerOut.output;
-            }
-        }
-        return null;
-    }
-
-    private static class StreamGobbler extends Thread {
-
-        private final InputStream stream;
-
-        private String output;
-
-        StreamGobbler(final InputStream stream) {
-            this.stream = stream;
-        }
-
-        @Override
-        public void run() {
-
-            try (BufferedReader bufferedReader = new BufferedReader(new InputStreamReader(stream))) {
-                output = bufferedReader.lines().collect(Collectors.joining());
-            } catch (IOException e) {
-                throw new RuntimeException("Problem reading from Vault Token Helper: " + e.getMessage(), e);
-            }
-        }
     }
 }
