+++ src/main/java/com/premiumminds/datagrip/vault/client/VaultTokenLoader.java
@@ -1,0 +1,9 @@
+package com.premiumminds.datagrip.vault.client;
+
+public interface VaultTokenLoader {
+
+    String get() throws Exception;
+
+}
+
+
